rework CompiledData to compile embedded refs.
This simplifies the process of finding references to other assets from within compiled data as well as the internals of CompiledData. We no longer have to provide a way to recurse though some data's type (reducing code duplication and generated code size). This comes at a cost to the size of the compiled data. We'll see if that winds up being a problem later.

[git-p4: depot-paths = "//enosi/ecs/": change = 467]