rework gfx::Renderer into gfx::RenderPass, most importantly moving ownership of Vulkan from it to the Editor for now. Also add vk.lh as a light interface to Vulkan which handles much of what Renderer used to.
We will probably move towards the Engine owning Vulkan once I start to setup building ecs as editor/game. For now it stays on Editor for simplicity. It will be important to move it out later though as we will want the debug system to be totally separate from whatever the Engine is running, so that it can display graphics (or be used via the terminal, in the case of the console controlling a headless server) and such even when the game or editor are in a broken state. 

[git-p4: depot-paths = "//enosi/ecs/": change = 597]