convert EntityEventBus to use linked lists allocated from a bump rather than a SLinkedList
This is primarily a compile-time save. Before this change, EntityEventBus.lpp.cpp would take ~15s to compile during a full build, and 8-9s when just compiling that file. After this change, it takes about 3s to compile during a full build, and 1.5s when just compiling this file. This also reduces the size of the sub lists struct from ~16k bytes to 2400 bytes. 

Not totally sure what kind of effect this will have on the game, but the way we were storing this kinda sucked. Too much templating for no reason.

[git-p4: depot-paths = "//enosi/ecs/": change = 672]