refactor iro::Array to store len, space, and allocator on itself rather than embedded before the array data
The way I originally implemented array wound up becoming more and more annoying over time, primarily when dealing with it in a debugger. It also doesn't seem to serve much of a purpose with how array data is dealt with in enosi. The point is that you are only passing around a pointer to refer to an array, but when we pass around array data we do so as Slices in 99% of cases.

[git-p4: depot-paths = "//enosi/ecs/": change = 686]