add unpacking api to turn binary data back into SourceData
unfortunately not going to use it right now, as I realized that it won't be very useful for map editing like I thought it would be. However we've needed this for awhile and it will likely be useful soon (eg. for outputting build metafiles more easily than we do now)

[git-p4: depot-paths = "//enosi/ecs/": change = 893]